Consolidate models into one file to make imports cleaner
