+++ server/models/__init__.py
@@ -1,3 +1,3 @@
-import server.models.login
-import server.models.transaction
-import server.models.user
+from server.models.login import Login
+from server.models.transaction import Transaction
+from server.models.user import User
