+++ server/routes/leaderboard.py
@@ -4,8 +4,7 @@ from flask import Blueprint
 from sqlalchemy.sql import func
 from server import db
 
-from server.models.transaction import Transaction
-from server.models.user import User
+from server.models import Transaction, User
 
 leaderboard_bp = Blueprint("leaderboard_bp", __name__)
